Extract CompartmentIdSplitter from FhirMappingProvider
Reduce FhirMappingProvider from 321 to 189 lines by extracting
compartment-related logic into a dedicated class.
